backend/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
spacy==3.7.2
numpy==1.24.3
chromadb==1.0.13
# Document Processing
pdfminer.six==20231228
python-docx==1.1.0
pydantic-settings==2.1.0
python-docx==1.1.0
# AI & ML - Updated versions for compatibility
sentence-transformers==2.7.0
huggingface_hub==0.20.3
transformers==4.36.2
scikit-learn==1.3.2
torch==2.1.0
anthropic==0.40.0
openai==1.93.0
# LangChain Dependencies - Compatible versions
langchain-core>=0.3.66
langchain-groq>=0.3.4
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
# Vector Database
pymilvus==2.3.4
faiss-cpu==1.7.4

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0
